+++ webots_ros2_msgs/msg/WbURDFRobot.msg
@@ -1,0 +1,4 @@
+string name
+string urdf_location
+string translation
+string rotation
