+++ webots_ros2_msgs/srv/SetWbURDFRobot.srv
@@ -1,0 +1,3 @@
+WbURDFRobot robot
+---
+bool success
