+++ webots_ros2_universal_robot/setup.py
@@ -14,6 +14,8 @@ data_files.append(('share/' + package_name + '/launch', [
     'launch/multirobot_launch.py',
     'launch/moveit_demo_launch.py',
     'launch/test_launch.py',
+    'launch/urdf_and_nodes_launch.py',
+    'launch/webots_launch.py',
 ]))
 data_files.append(('share/' + package_name + '/worlds', [
     'worlds/universal_robot.wbt',
@@ -59,8 +61,7 @@ setup(
         'console_scripts': [
             'ur5e_controller = webots_ros2_universal_robot.ur5e_controller:main',
             'abb_controller = webots_ros2_universal_robot.abb_controller:main',
-            'supervisor_driver = webots_ros2_universal_robot.supervisor_driver:main',
-            'supervisor_node = webots_ros2_universal_robot.supervisor_node:main'
+            'supervisor_spawner = webots_ros2_universal_robot.supervisor_spawner:main',
         ]
     }
 )
